Optional improved support for amending git commits
Add opt-in feature to allow the diff and status to be extracted from the
commit message template if the verbose option is used with git commit,
e.g. git commit --verbose or git config --global commit.verbose=true

When amending commits, this allows the author of the commit message to
see all of the changes, in both the status and diff windows, if they
enabled the verbose option for git commit. Without the verbose option,
the behaviour is as before, show only new changes, if there are any.

Enable in your vimrc with:

    let g:committia#git#use_verbose = 1