Respect git comment char in status window
Prefix lines in status with git comment char rather than #

For most people, who accept the default comment char of #, this makes
absolutely no difference. However, if you do set core.commentChar to
something other than the default, this change means that character is
now consistently used in the status window, whether it is populated by
by running `git status` or extracted from the verbose commit message.

Technical note: this assumes that the last line in the message before
the diff is a comment, which it should be unless the --no-status option
is used, in which case it falls back to using '#' rather than execute a
git command. The alternative would be to run `git config core.commentChar`